feat(break-change): Split files
