+++ src/controller/breakingChanges/utils.ts
@@ -3,6 +3,22 @@ import { KeyHandler } from '../../redis/key_handler';
 import { parseInputDate } from '../../graphql/resolvers/getOperationUsageTrack';
 import { RedisRepository } from '../../redis/redis';
 import { getTimestamp } from '../../redis/utils';
+import {OperationChange} from "./operation";
+import {EnumChange} from "./enum";
+import {FieldChange} from "./field";
+import {TypeChange} from "./type";
+import {Change} from "@graphql-inspector/core";
+
+type BreakingChangeType = OperationChange | EnumChange | FieldChange | TypeChange;
+
+export const getBreakingChangesTypes = (): BreakingChangeType[] => {
+	return [
+		new FieldChange(),
+		new TypeChange(),
+		new OperationChange(),
+		new EnumChange(),
+	];
+}
 
 export const checkUsage = async (
 	operations: ClientOperationsDTO,
@@ -34,3 +50,7 @@ export const checkUsage = async (
 		return (acc += cur.total);
 	}, 0);
 };
+
+export const validateBreakingChange = (types: string[] ,change: Change): boolean => {
+	return types.includes(change.type);
+}
